// apps/web/src/modules/documents/domain/enums/toolbar-actions.enum.ts
export enum ToolbarAction {
  HEADING_1 = 'heading_1',
  HEADING_2 = 'heading_2',
  HEADING_3 = 'heading_3',
  HEADING_4 = 'heading_4',
  HEADING_5 = 'heading_5',
  HEADING_6 = 'heading_6',
  BOLD = 'bold',
  ITALIC = 'italic',
  STRIKETHROUGH = 'strikethrough',
  BULLET_LIST = 'bullet_list',
  NUMBERED_LIST = 'numbered_list',
  CHECKLIST = 'checklist',
  LINK = 'link',
  IMAGE = 'image',
  CODE = 'code',
  TABLE = 'table',
  BLOCKQUOTE = 'blockquote',
  HORIZONTAL_RULE = 'horizontal_rule',
}

export enum ToolbarGroup {
  TEXT = 'text',
  LISTS = 'lists',
  INSERT = 'insert',
  BLOCK = 'block',
}